requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python-headless>=4.8.0
fpdf2>=2.7.6
numpy>=1.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
